Update utils.py
fixed the indexing (similar to what it was done in commit 490c85c). the original indices `4` and `5` are not consistent with the directory structure that is defined in the README, it should be `3` and `4`. But I used the indices `-1` and `-2` so that the code would require less changes in the future if the directory structure changes.